Proyecto-2/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
cryptography==47.0.0
python-dotenv==1.2.2
python-multipart==0.0.27